Move sgx specific properties from bit/properties.h
Moved sgx specific properties into bits/sgx/sgxproperties.h
